Codec.JSON.Decoder.Stream

Description

Functions for parsing JSON with incremental output.

Stream

data Stream a (m :: Type -> Type) r

A list parametrized by an element type a, an effect type m and the return type r.

Constructors

Yield a (Stream a m r)

Effect (m (Stream a m r))

Return r

Run

stream :: Source a r -> Blank -> Stream a Partial (Blank, Either (Path, Error) r)

Run a stream incrementally.

Source

data Source a r

Decoder type for streams.

mapSource :: (a -> b) -> Source a r -> Source b r

Apply a function to each element of the stream-decoder.

foldSource :: (b -> a -> b) -> (b -> r -> b) -> b -> Source a r -> Decoder b

Left-associative fold of a stream-decoder, lazy in the accumulator.

foldSource' :: (b -> a -> b) -> (b -> r -> b) -> b -> Source a r -> Decoder b

Left-associative fold of a stream-decoder, strict in the accumulator.

sourceDecoder :: Decoder a -> r -> Source a r

Use a non-incremental decoder as a single-element stream.

Object

Key

data KeyDecoder a

Object pair name decoder type

stringKey :: KeyDecoder String

Decode a JSON object pair name as a String.

textKey :: KeyDecoder Text

Decode a JSON object pair name as a strict Text.

lazyTextKey :: KeyDecoder Text

Decode a JSON object pair name as a lazy Text.

copyKey :: KeyDecoder a -> KeyDecoder (JSONKey, a)

Get a copy of the underlying input alongside the object pair name.

Sources

sourceObject :: KeyDecoder k -> (k -> r -> Source a r) -> r -> Source a r

Stream a JSON object. Object pair names are parsed with the given KeyDecoder.

sourceObject_ :: (JSONKey -> r -> Source a r) -> r -> Source a r

Stream a JSON object. Object pair names are copied verbatim.

Array

sourceArray :: (Word -> r -> Source a r) -> r -> Source a r

Stream a JSON array. The Word argument is the element index, starts at 0.