Codec.JSON.Decoder.TH

Description

Functions for precompiling decoders.

Type classes

Functor

class FunctorQ (f :: Type -> Type) where

Variation of Functor that uses typed splices.

Methods

fmapQ :: Code Q (a -> b) -> f a -> f b

(|$|) :: FunctorQ f => Code Q (a -> b) -> f a -> f b infixl 4

Variation of (<$>) that uses typed splices.

Applicative

class FunctorQ f => ApplyQ (f :: Type -> Type) where

Variation of Apply that uses typed splices.

Methods

(|*|) :: f (a -> b) -> f a -> f b infixl 4

class ApplyQ f => ApplicativeQ (f :: Type -> Type) where

Variation of Applicative that uses typed splices.

Methods

pureQ :: Code Q a -> f a

Decode

Object

class PairQ (f :: Type -> Type)

Class of precompiled JSON object parsers.

Minimal complete definition

pairQ, pairMaybeQ

(.|) :: PairQ f => Text -> Code Q (Decoder a) -> f a

Decode a JSON value at the given key.

Fails if the key is not present in the JSON object.

(?|) :: PairQ f => Text -> Code Q (Decoder a) -> f (Maybe a)

Decode an optional JSON value at the given key.

Returns Nothing if the key is not present in the JSON object or if the value is a JSON null.

Applicative

data PairsQ a

Precompiled applicative object parser.

pairsQ :: PairsQ a -> Code Q (Decoder a)

Parse a JSON object as a tuple of key/value pairs.

In presence of duplicate keys in the supplied JSON object prefers earlier entries.

Array

class ElementQ (f :: Type -> Type) where

Class of precompiled JSON array parsers.

Methods

elementQ :: Code Q (Decoder a) -> f a

Decode a JSON value at the current position.

Applicative

data ElementsQ a

Precompiled applicative array parser.

elementsQ :: ElementsQ a -> Code Q (Decoder a)

Decode a JSON array as a non-empty tuple.

Fails if the JSON array holds a different number of elements.