Codec.JSON.Decoder.Unsafe

Description

Data structure internals, helper operations and unsafe functions.

Itself

newtype Decoder a

The decoder type, parametrized by the return type a.

Constructors

Decoder

Fields

runDecoder :: Path -> Bitmask -> K -> Parser (Path, Error) a

data Path

Subset of JSONPath used for error reporting, represented as a snoc-list.

Constructors

Key
Object pair.

Fields

!Path

!JSONKey

Name of the pair.

Index
Array element.

Fields

!Path

!Word

Index of the element. Starts at 0.

Root
Root node identifier.

newtype Bitmask

Set of six JSON value types that can be tried by any particular parser.

Constructors

Bitmask Word8

Bundled Patterns

pattern EmptyBitmask :: Bitmask

pattern ObjectBit :: Bitmask

pattern ArrayBit :: Bitmask

pattern NumberBit :: Bitmask

pattern StringBit :: Bitmask

pattern BooleanBit :: Bitmask

pattern NullBit :: Bitmask

data K

First byte of a JSON value.

Constructors

O
Object.

A
Array.

N
Number.

S
String.

T
True.

F
False.

X
Null.

data Gravity

Whether the error can be recovered from.

Constructors

Benign
Personal judgement.

Fatal
JSON data itself is malformed.

data Error

Kinds of errors known to the parser.

Constructors

Mismatch
Parser does not serve this JSON type.

Fields

!Bitmask

Set of JSON types that were tried.

!K

JSON type in question.

Malformed
Input could not be converted properly.

Fields

!Gravity

!String

Error description.

AbruptEnd
Input ended abruptly.

Decode

Object

radixTree :: Decoder (RadixTree JSON)

Decode a JSON object as a RadixTree of opaque JSON values.

Keys are guaranteed to be encoded as well-formed UTF-8.

In presence of duplicate keys in the supplied JSON object prefers later entries.

Surrogate Unicode code points in JSON strings are treated as valid scalar values.

radixTree' :: Decoder (RadixTree JSON)

Decode a JSON string as a RadixTree of opaque JSON values.

Keys are guaranteed to be encoded as well-formed UTF-8.

In presence of duplicate keys in the supplied JSON object prefers later entries.

Surrogate Unicode code points in JSON strings are treated as invalid input.

class Pair (f :: Type -> Type) where

Class of JSON object parsers.

Methods

pair :: Text -> Decoder a -> f a

Ditto (.:).

pairMaybe :: Text -> Decoder (Maybe a) -> f (Maybe a)

Decode an optional JSON value at the given key.

Returns Nothing if the key is not present in the supplied JSON object.

class PairQ (f :: Type -> Type) where

Class of precompiled JSON object parsers.

Methods

pairQ :: Text -> Code Q (Decoder a) -> f a

Ditto (.|).

pairMaybeQ :: Text -> Code Q (Decoder (Maybe a)) -> f (Maybe a)

Decode an optional JSON value at the given key.

Returns Nothing if the key is not present in the supplied JSON object.

Array

snocList :: Decoder a -> Decoder (Word, [a])

Decode a JSON array as a snoc-list of values, alongside its length.

Number

bignum :: Decoder (Integer, Integer)

Decode a JSON number as a two-tuple of decimal significand and decimal exponent.

String

unsafeString :: Decoder String

Decode a JSON string as a String.

Surrogate Unicode code points are treated as valid scalar values.

unsafeText :: Decoder Text

Decode a JSON text as a strict Text.

Surrogate Unicode code points are treated as valid scalar values.

unsafeLazyText :: Decoder Text

Decode a JSON string as a lazy Text.

Surrogate Unicode code points are treated as valid scalar values.

Stream

newtype Source a r

Decoder type for streams.

Constructors

Source

Fields

runSource :: Path -> Bitmask -> K -> Stream a (Parser (Path, Error)) r

newtype KeyDecoder a

Object pair name decoder type

Constructors

KeyDecoder (Decoder a)