Codec.JSON.Encoder

Description

Pure incremental RFC 8259 JSON encoder.

Encoding functions in this module produce JSON without insignificant whitespace.

Itself

data Encoder

The encoder type, wrapping a well-formed raw JSON value.

Run

encode :: Encoder -> Builder

Convert any encoded value into a Builder.

Encode

Object

data KeyEncoder

Object pair name encoder type.

stringKey :: String -> KeyEncoder

Encode a String as a JSON object pair name.

textKey :: Text -> KeyEncoder

Encode a strict Text as a JSON object pair name.

lazyTextKey :: Text -> KeyEncoder

Encode a lazy Text as a JSON object pair name.

jsonKey :: JSONKey -> KeyEncoder

Use a raw object pair name as a JSON object pair name.

Semigroup

emptyObject :: Encoder

Empty JSON object.

object1 :: Object1 -> Encoder

Encode a non-empty JSON object.

data Object1

Non-empty JSON object encoder type.

pair :: KeyEncoder -> Encoder -> Object1

Use a JSON object pair name and a JSON value as a non-empty JSON object pair.

List

data Pair

JSON object encoder type.

(.=) :: KeyEncoder -> Encoder -> Pair

Use a JSON object pair name and a JSON value as a JSON object pair.

(?=) :: KeyEncoder -> Maybe Encoder -> Pair

Use a JSON object pair name and an optional JSON value as a JSON object pair.

Providing Nothing excludes the pair from the resulting object.

pairs :: [Pair] -> Encoder

Encode a list of Pairs as a JSON object.

Array

Semigroup

emptyArray :: Encoder

Empty JSON array.

array1 :: Array1 -> Encoder

Encode a non-empty JSON array.

data Array1

Non-empty JSON array encoder type.

element :: Encoder -> Array1

Use a JSON value as a non-empty JSON array element.

List

list :: [Encoder] -> Encoder

Encode a list of encoders as a JSON array.

nonEmpty :: NonEmpty Encoder -> Encoder

Encode a non-empty list of encoders as a JSON array.

Number

Unsigned integer

word8 :: Word8 -> Encoder

Encode a Word8 as a JSON number.

word16 :: Word16 -> Encoder

Encode a Word16 as a JSON number.

word32 :: Word32 -> Encoder

Encode a Word32 as a JSON number.

word64 :: Word64 -> Encoder

Encode a Word64 as a JSON number.

word :: Word -> Encoder

Encode a Word as a JSON number.

natural :: Natural -> Encoder

Encode a Natural as a JSON number.

Signed integer

int8 :: Int8 -> Encoder

Encode an Int8 as a JSON number.

int16 :: Int16 -> Encoder

Encode an Int16 as a JSON number.

int32 :: Int32 -> Encoder

Encode an Int32 as a JSON number.

int64 :: Int64 -> Encoder

Encode an Int64 as a JSON number.

int :: Int -> Encoder

Encode an Int as a JSON number.

integer :: Integer -> Encoder

Encode an Integer as a JSON number.

Floating-point

float :: Float -> Encoder

Encode a Float as a JSON number.

NaN, \(+\infty\) and \(-\infty\) are instead encoded as "NaN", "Infinity" and "-Infinity" JSON strings respectively.

double :: Double -> Encoder

Encode a Float as a JSON number.

NaN, \(+\infty\) and \(-\infty\) are instead encoded as "NaN", "Infinity" and "-Infinity" JSON strings respectively.

String

string :: String -> Encoder

Encode a String as a JSON string.

text :: Text -> Encoder

Encode a strict Text as a JSON string.

lazyText :: Text -> Encoder

Encode a lazy Text as a JSON string.

Boolean

bool :: Bool -> Encoder

Encode a Bool as a JSON boolean.

Null

maybe :: Maybe Encoder -> Encoder

Encode a Maybe, using null as Nothing.

null :: Encoder

JSON null.

Raw

data JSON

Raw JSON value.

json :: JSON -> Encoder

Use a raw value as an encoder.